requirements.txt
fastapi
uvicorn
python-dotenv
streamlit
openai
pydantic 
pydantic-settings
langchain
langchain-core
langchain-openai
langchain-community